Move helper functions out of class definition
These seem like they belong elsewhere
